@@ -1,25 +1,24 @@
 using Microsoft.AspNetCore.Mvc;
-using Expense.Models;
-using System.Threading.Tasks;
+using Swashbuckle.AspNetCore.Annotations;
+using expense.Models;
+namespace expense.Controllers;
 
-namespace expense.Controllers
+[Route("api")]
+[ApiController]
+[Tags("Expense")]
+public class VersionController : ControllerBase
 {
-    [Route("api")]
-    [ApiController]
-    public class VersionController : ControllerBase
+
+    private readonly IVersionContext _context;
+
+    public VersionController(IVersionContext context)
     {
+        _context = context;
+    }
 
-        private readonly IVersionContext _context;
-
-        public VersionController(IVersionContext context)
-        {
-            _context = context;
-        }
-
-        [HttpGet]
-        public async Task<ActionResult<string>> Version()
-        {
-            return await _context.GetVersion();
-        }
+    [HttpGet]
+    public async Task<ActionResult<string>> Version()
+    {
+        return await _context.GetVersion();
     }
 }
